requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
requests==2.26.0
celery==5.2.7
python-dotenv==0.20.0
psycopg2-binary==2.9.3
django-redis==3.1.6
django-cacheops==6.0
langdetect==1.0.9
pytest-django==4.5.2
pytest-factoryboy==2.5